[DEVELOP][ADD] 添加 保存商品的接口(还缺少保存标题图片和详情图片)

+++ src/main/java/com/jichuangtech/clothshopserver/repository/ColorRepository.java
@@ -7,6 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Set;
 
-public interface ColorRepository extends JpaRepository<ColorEntity, Long>{
-
+public interface ColorRepository extends JpaRepository<ColorEntity, Integer>{
 }
